net: Implement random ethaddr fallback in eth.c
Commit bef1014b31c5b33052bcaa865ba3618d73e906f0 upstream.

Backport changes CONFIG_NET_RANDOM_ETHADDR to CONFIG_RANDOM_MACADDR,
net_random_ethaddr to eth_random_enetaddr, and is_zero_ethaddr to
is_zero_ether_addr, and keeps existing function order.

Upstream log:

Implement the random ethaddr fallback in eth.c so it is in a common
place and not reimplemented in each board or driver that wants this
behavior.

Signed-off-by: Joe Hershberger <joe.hershberger@ni.com>
Reviewed-by: Simon Glass <sjg@chromium.org>

Signed-off-by: Milton D. Miller II <miltonm@us.ibm.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>