drivers/net/aspeednic: Fix unreliable networking
Hardware must be told to move data from its RX FIFO to system memory where
the driver can then consume it.  There are two mechanisms for this, a 'do
this automatically' register which can be set on device init or simply a
'copy data now' polling style which is what this patch does.

Both options were investigated and as the entire network stack of u-boot is
polling and we're only polling when we expect to receive a message it make
sense to continue this down to the hardware level as well, doing the
automatic copy actually causes more noise and unnecessary work.

It should be noted that the NIC was largely functional before this patch.
It appears that the reason it was working was that the hardware would quite
happily DMA into the receive ring as long as there was was spare space.
During heavy loads or on busy networks the receiving ring can fill up which
causes the hardware to stop (and never resume) its DMA activities unless
explicitly told to.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>