ARM: ast2400: Move UART setup outside DRAM init
When a BMC reboot occurs an SOC reset is performed.  An SOC reset does
not reset the DRAM.  The uboot platform code initialized UART in dram
init section.  So an SOC reset was causing the UART to not be
initialized on reboot.  This moves the UART setup code before dram init
stage.  This change should be integrated if SOC reset method is used to
reboot BMC otherwise you won't see a u-boot console.

Signed-off-by: Norman James <nkskjames@gmail.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>